Some more preparation code for gameplay

+++ src/model/Memory.h
@@ -40,7 +40,7 @@ public:
     
     Event event;
     
-    int when;
+    long when;
     
     Character *who;
